--- src/main/java/ru/mirea/megatracker/api/PriceInfoRUB.java
@@ -1,9 +1,0 @@
-package ru.mirea.megatracker.api;
-
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class PriceInfoRUB extends PriceInfo {
-
-}
