+++ src/main/java/ru/mirea/megatracker/api/response/ExchangePairApiResponse.java
@@ -1,4 +1,4 @@
-package ru.mirea.megatracker.api;
+package ru.mirea.megatracker.api.response;
 
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
@@ -7,6 +7,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import ru.mirea.megatracker.api.CoinInfo;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 @Getter
